Fix import side effect
