+++ .github/workflows/build-and-test.yml
@@ -73,7 +73,6 @@ jobs:
           TD_COMPANY_ELASTICSEARCH_INDEX: "stocketablissement-production"
           TD_COMPANY_ELASTICSEARCH_URL: ${{ secrets.TD_COMPANY_ELASTICSEARCH_URL }}
           TD_COMPANY_ELASTICSEARCH_CACERT: ${{ secrets.TD_COMPANY_ELASTICSEARCH_CACERT }}
-          ELASTIC_SEARCH_URL: "http://elasticsearch"
 
   doc:
     runs-on: ubuntu-latest
